requirements.txt
# 工具插件依赖
requests>=2.28.0
beautifulsoup4>=4.11.0
pyyaml>=6.0

# 异步支持
aiohttp>=3.8.3

# 日志和工具
python-json-logger>=2.0.4 